- STARTED reworking ResourceController. - STARTED reworking Structure and its sublasses, mainly focusing on giving Windmill temporary logic until Villagers and Task system are more accommodating. - Windmill has a list of requirements for Leveling it up. If it meets those resource requirements when it turns night, it will consume those resources and Level up. - Probably removing Resource classes in future, as they don't seem to be actual tangible objects in the game with any unique functions other than seeing how many of each we have to do tasks with.

+++ Assets/Project/Scripts/GameControllers/StructureController.cs
@@ -41,7 +41,7 @@ public class StructureController : MonoBehaviour
 	#endregion
 	
 	#region Create Structure(s)
-	public void CreateNewStructure(string name, float age, Sprite icon, Vector3 pos)
+	public void CreateNewStructure(string name, int age, Sprite icon, Vector3 pos)
 	{
 		GameObject instance = Instantiate (Resources.Load ("Prefabs/Structures/StructureGO")) as GameObject;
 		Structure structureInstance = instance.GetComponent<Structure>();
